+++ CHANGELOG.md
@@ -4,8 +4,15 @@ All notable changes to `laravel-shopping-cart` will be documented in this file.
 
 # v1.0.1 (2017-11-07)
 
-### Fixed
+### Changed
+- Cart::empty() -> Cart::clear()
+- Cart::emptyConditions() -> Cart::clearConditions()
+- README.md - Changed respectively.
+
+### Added
 - README.md - Description for setConditionAsArray added for Cart and Item.
+
+### Fixed
 - phpunit
 
 # v1.0.0 (2017-11-06)